+++ src/test/java/pairmatching/domain/CrewTest.java
@@ -1,11 +1,9 @@
 package pairmatching.domain;
 
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
-import pairmatching.Course;
+import pairmatching.domain.program.Course;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class CrewTest {
